requirements.txt
opencv-python==4.8.1.78
numpy>=1.24.3
scikit-learn==1.3.2
tensorflow-cpu>=2.15.0
streamlit==1.28.1
Pillow==10.1.0
h5py>=3.9.0

